gae_dash/requirements.txt
asn1crypto==0.24.0
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.3
djangorestframework==3.9.0
google-api-core==1.5.2
google-auth==1.6.0
google-cloud-monitoring==0.30.1
googleapis-common-protos==1.5.5
grpcio==1.16.0
idna==2.7
mysqlclient==1.3.13
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
